Move Twitter Bot to its own directory
