+++ src/scripts/tweet.js
@@ -1,6 +1,6 @@
 import { Logger } from 'blocks/utilities/logger';
-import { Screenshot } from 'blocks/utilities/screenshot';
-import { TwitterBot } from 'blocks/utilities/twitterBot';
+import { Screenshot } from 'blocks/twitterBot/screenshot';
+import { TwitterBot } from 'blocks/twitterBot';
 
 export const tweet = async () => {
   Logger.log('Twitter bot is starting up...\n');
